refactor(tests): centralize immutable config fixtures
