+++ src/test-kernel/agent/modelHandlers/GoogleInteractionsBackground.vitest.ts
@@ -2,15 +2,13 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { ModelProvider } from 'llm-zoo';
 
-// Local imports - agent and config
+// Local imports - test support, agent, and config
+import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import { ModelHandlerGoogleInteractions } from '@agent/modelHandlers/google/modelHandlerGoogleInteractions';
 import { noopTrace } from '@agent/trace/noopTrace';
 import * as configModule from '@utils/config/configUtils';
 import * as providerConfigModule from '@utils/config/providerConfig';
-
-// Local imports - test fixtures
-import { buildTestModelConfig } from './testFixtures';
 
 // Type imports
 import type { Interactions } from '@google/genai';
