+++ src/test-kernel/agent/modelHandlers/GoogleInteractionsChaining.vitest.ts
@@ -2,14 +2,12 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { ModelProvider } from 'llm-zoo';
 
-// Local imports - agent and config
+// Local imports - test support, agent, and config
+import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import { ModelHandlerGoogleInteractions } from '@agent/modelHandlers/google/modelHandlerGoogleInteractions';
 import { noopTrace } from '@agent/trace/noopTrace';
 import * as configModule from '@utils/config/configUtils';
-
-// Local imports - test fixtures
-import { buildTestModelConfig } from './testFixtures';
 
 // Type imports
 import type { Interactions } from '@google/genai';
