+++ src/test-kernel/agent/modelHandlers/ModelHandlerOpenAINormalize.vitest.ts
@@ -3,14 +3,12 @@ import { strict as assert } from 'node:assert';
 import { describe, it } from 'vitest';
 import { ModelProvider } from 'llm-zoo';
 
-// Local imports - agent
+// Local imports - test support and agent
+import { buildTestModelConfig } from '@test/support/modelConfigTestUtils';
 import { ModelHandlerDashScope } from '@agent/modelHandlers/openai/modelHandlerDashScope';
 import { ModelHandlerDeepSeek } from '@agent/modelHandlers/openai/modelHandlerDeepSeek';
 import { ModelHandlerKimi } from '@agent/modelHandlers/openai/modelHandlerKimi';
 import { noopTrace } from '@agent/trace/noopTrace';
-
-// Local imports - test fixtures
-import { buildTestModelConfig } from './testFixtures';
 
 function createClientStub() {
   const createCalls: any[] = [];
